All Website structure, Assistance Service Page, Category Pages and Who We Are
This commit include the following changes:
- All pages that will be on the website have been created, with the
header, the right orientation info, a pseudocontent and a pseudofooter
(to be done later);
- The pages devices, SL service, Assistance services now contain a
slider to redirect to the specific category
- A tab panel structure has been introduced in Who We Are (content to be
filled)
- The page Assistance Service now include a left tab panel (completely
working) and information are taken from the new table in the database.
In particular, a specific structure of JSON has been decided for the
body field.